refactor: dragover, drop 이벤트도 루트 노드에 등록
